Update README to describe usage in non-admin views
Thank you for taking a look at my PR!

I wanted to use this on a non-admin view and ran into a gotcha around needing to include a `{{ form.media }}` tag in my template. I'm proposing an update to the README in case anyone else runs into the same trouble.